Update dependency namespace
