+++ core/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.kodabots.sdk.core">
+    package="ai.koda.mobile.sdk.core">
     <uses-permission android:name="android.permission.INTERNET" />
 </manifest>
