+++ core/src/main/java/ai/koda/mobile/sdk/core/Extensions.kt
@@ -1,4 +1,4 @@
-package com.kodabots.sdk.core
+package ai.koda.mobile.sdk.core
 
 import android.util.Log
 import android.webkit.WebView
