+++ core/src/main/java/ai/koda/mobile/sdk/core/KodaBotsConfig.kt
@@ -1,4 +1,4 @@
-package com.kodabots.sdk.core
+package ai.koda.mobile.sdk.core
 
 import android.graphics.Typeface
 import android.graphics.drawable.Drawable
