+++ core/src/main/java/ai/koda/mobile/sdk/core/KodaBotsPreferences.kt
@@ -1,4 +1,4 @@
-package com.kodabots.sdk.core
+package ai.koda.mobile.sdk.core
 
 import android.content.Context
 import android.content.SharedPreferences
