+++ core/src/main/java/ai/koda/mobile/sdk/core/KodaBotsWebViewClient.kt
@@ -1,4 +1,4 @@
-package com.kodabots.sdk.core
+package ai.koda.mobile.sdk.core
 
 import android.webkit.WebResourceRequest
 import android.webkit.WebView
