+++ gradle/libs.versions.toml
@@ -1,5 +1,6 @@
 [versions]
 artifactregistry-gradle-plugin-version = "2.2.1"
+koda-core = "1.2.0"
 kotlin = "2.0.0"
 core-ktx = "1.13.1"
 appcompat = "1.7.0"
@@ -20,6 +21,8 @@ security-crypto = "1.1.0-alpha06"
 
 [libraries]
 androidx-security-crypto = { module = "androidx.security:security-crypto", version.ref = "security-crypto" }
+koda-core = { module = "ai.koda.mobile.sdk:koda-core", version.ref = "koda-core" }
+koda-core-staging = { module = "ai.koda.mobile.sdk:koda-core-staging", version.ref = "koda-core" }
 kotlin-stdlib = { group = "org.jetbrains.kotlin", name = "kotlin-stdlib", version.ref = "kotlin" }
 androidx-core = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
 androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
