+++ sample/src/debug/AndroidManifest.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.kodabots.sdk.sample">
+    package="ai.koda.mobile.sdk.sample">
 
     <application
-        android:name=".SampleApplication"
+        android:name="ai.koda.mobile.sdk.sample.SampleApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name="ai.koda.mobile.sdk.sample.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,7 +19,7 @@
             </intent-filter>
         </activity>
         <meta-data
-            android:name="com.kodabots.sdk.ClientToken"
+            android:name="ai.koda.mobile.sdk.ClientToken"
             android:value="${KODA_AI_CLIENT_TOKEN}" />
     </application>
