+++ sample/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/activity_main_root"
-    tools:context=".MainActivity">
+    tools:context="ai.koda.mobile.sdk.sample.MainActivity">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/activity_main_controls_expander"
