+++ sample/src/main/res/values/colors.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#e0e0e0</color>
+    <color name="colorPrimary">#555555</color>
     <color name="colorPrimaryLight">#ffffff</color>
-    <color name="colorPrimaryDark">#aeaeae</color>
+    <color name="colorPrimaryDark">#515151</color>
     <color name="colorAccent">#3f51b5</color>
     <color name="colorAccentLight">#757de8</color>
     <color name="colorAccentDark">#002984</color>
